+++ changelog.md
@@ -8,14 +8,17 @@ added new styles:
 * classic
 commands changed:
 * version command now has a branch, edition and isExperimental shown up
+* help now has --regex option
 added:
 * new command return type expectedError
 * new entries for styles
 * config, colors and styles were added to this in commands
 * terminalApi has now styles and colors getters
 * verApi has now a branch, edition and isExperimental
+* filter in command Collection
 fixed:
 * multiline commands are now properly supported as new typing system was introduced
+* wrongly written inputbox in logs
 
 compatibility:
 * cursor is now hidden by a default